# Speed up browser tests