I noticed that these tests are running really slow and wanted to see if I could make them faster. A few removals of some waiting calls and they are running faster (on my machine).

## Before

```
    ✔ should bubble when bubbles is true (2063ms)
    ✔ should hot replace a module (2074ms)
    ✔ should hot replace a bubbled module (3073ms)
    ✔ hot replaces multiple bubbled modules (3109ms)
    ✔ reloads the page when a module has no hot replacable parent (4091ms)


  5 passing (15s)
```

## After

```
    ✔ should bubble when bubbles is true (2058ms)
    ✔ should hot replace a module (81ms)
    ✔ should hot replace a bubbled module (73ms)
    ✔ hot replaces multiple bubbled modules (68ms)
    ✔ reloads the page when a module has no hot replacable parent (90ms)


  5 passing (3s)
```